+++ backend/src/controllers/game.controller.ts
@@ -1,7 +1,13 @@
 import {NextFunction, Request, Response} from "express"
 import createError from "http-errors";
-import {createGameSchema, CreateGameType} from "../validation/game.validation";
-import {checkPlayerGameEligibility, createGame, checkPlayerListUnique} from "../services/game.service";
+import {createGameSchema} from "../validation/game.validation";
+import {
+    checkPlayerGameEligibility,
+    createGame,
+    checkPlayerListUnique,
+    getDefaultRound,
+    getWind, getGame
+} from "../services/game.service";
 import {Game, Player} from "@prisma/client";
 import {findPlayerByUsernames} from "../services/player.service";
 import {getCurrentSeason} from "../services/season.service";
@@ -14,25 +20,44 @@ const getGamesHandler = async (req: Request, res: Response): Promise<void> => {
     }
 }
 
-const getGameHandler = async (req: Request, res: Response): Promise<void> => {
-    try {
-        res.status(200).json({})
-    } catch (error) {
-        throw error
+const getGameHandler = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+    const id = Number(req.params.id)
+    if (isNaN(id)) {
+        next(createError.NotFound("Game not found"))
+        return
     }
+
+    getGame(id).then((newGame) => {
+        if (!newGame) {
+            next(createError.NotFound("Game not found"))
+            return
+        }
+
+        res.status(200).json({
+            id: newGame.id,
+            gameType: newGame.gameType,
+            gameVariant: newGame.gameVariant,
+            status: newGame.status,
+            recordedById: newGame.recordedById,
+            players: newGame.players,
+            rounds: newGame.japaneseRounds || newGame.hongKongRounds
+        })
+    }).catch((err) => {
+        next(createError.InternalServerError(err.message))
+    })
 }
 
 const createGameHandler = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     try {
-        const createGameObject: CreateGameType = await createGameSchema.validate(req.body)
+        const {players, gameType, gameVariant} = await createGameSchema.validate(req.body)
 
-        checkPlayerListUnique(createGameObject.players)
-        checkPlayerGameEligibility(createGameObject.gameVariant, req.player)
-        const playerList = await findPlayerByUsernames(createGameObject.players);
-        playerList.forEach((player) => checkPlayerGameEligibility(createGameObject.gameVariant, player))
-        const playersQuery = playerList.map((player: Player, index) => {
+        checkPlayerListUnique(players)
+        checkPlayerGameEligibility(gameVariant, req.player)
+        const playerList = await findPlayerByUsernames(players);
+        playerList.forEach((player) => checkPlayerGameEligibility(gameVariant, player))
+        const playersQuery = playerList.map((player: Player) => {
             return {
-                wind: getWind(createGameObject.players.indexOf(player.username)),
+                wind: getWind(players.indexOf(player.username)),
                 player: {
                     connect: {
                         id: player.id
@@ -43,33 +68,15 @@ const createGameHandler = async (req: Request, res: Response, next: NextFunction
 
         const season = await getCurrentSeason();
 
-        const newGame: Game = await createGame(createGameObject, playersQuery, req.player.id, season.id)
+        const defaultRoundQuery = getDefaultRound(gameVariant);
+
+        const newGame: Game = await createGame(gameVariant, gameType, playersQuery, defaultRoundQuery, req.player.id, season.id)
 
         res.status(201).json({
-            gameId: newGame.id,
-            rounds: [{
-                roundCount: 1,
-                roundNumber: 1,
-                roundWind: "EAST"
-            }]
+            id: newGame.id
         })
     } catch (error: any) {
         next(createError.BadRequest(error.message))
-    }
-}
-
-const getWind = (index: number): string => {
-    switch (index) {
-        case 0:
-            return "EAST"
-        case 1:
-            return "SOUTH"
-        case 2:
-            return "WEST"
-        case 3:
-            return "NORTH"
-        default:
-            return "NONE"
     }
 }